metab_server/test/requirements.txt
MouseInfo==0.1.3
numpy==2.2.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytweening==1.2.0
setuptools==78.1.1
websockets==15.0.1
wheel==0.45.1